Windows, "default values" and a lot of modifications.
I added the wxWidgets and wxc.dll needed for windows to a folder mswin
in the xtras folder. They need to be found by the apps so moving it to
c:\windows is a good way to do that.

The example now uses a "wxPanel" inside the "wxFrame" which is the
prefered way. Most controls should be inside a wxPanel to support Tab
traversal and other stuff which is not implemented in wxFrame.